LayoutTests/fast/encoding/char-decoding-expected.txt

This tests decoding characters in various character sets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS decode('UTF-8', '%E2%88%9A') is 'U+221A'
PASS decode('macintosh', '%C3') is 'U+221A'
PASS decode('MacRoman', '%C3') is 'U+221A'
PASS decode('gb2312', '%A3%A0') is 'U+3000'
PASS decode('gb_2312-80', '%A3%A0') is 'U+3000'
PASS decode('chinese', '%A3%A0') is 'U+3000'
PASS decode('gbk', '%A3%A0') is 'U+3000'
PASS decode('gb18030', '%A3%A0') is 'U+3000'
PASS decode('korean', '%A2%E6') is 'U+20AC'
PASS decode('korean', '%A1%A4') is 'U+00B7'
PASS decode('korean', '%A1%A9') is 'U+00AD'
PASS decode('korean', '%A1%AA') is 'U+2015'
PASS decode('korean', '%A1%AD') is 'U+223C'
PASS decode('korean', '%A2%A6') is 'U+FF5E'
PASS decode('korean', '%A2%C1') is 'U+2299'
PASS decode('korean', '%1A') is 'U+001A'
PASS decode('korean', '%1C') is 'U+001C'
PASS decode('korean', '%8F%A1') is 'U+B8EA'
PASS decode('korean', '%B4%D3') is 'U+B2D2'
PASS decode('korean', '%A2%41') is 'U+C910'
PASS decode('euc-kr', '%A2%41') is 'U+C910'
PASS decode('korean', '%A2%41') is 'U+C910'
PASS decode('windows-949', '%A2%41') is 'U+C910'
PASS decode('iso-ir-149', '%A2%41') is 'U+C910'
PASS decode('KS_C_5601-1987', '%A2%41') is 'U+C910'
PASS decode('KS_C_5601-1989', '%A2%41') is 'U+C910'
PASS decode('iso-8859-9', '%80') is 'U+20AC'
PASS decode('iso-8859-9', '%9F') is 'U+0178'
PASS decode('iso-8859-9', '%FD') is 'U+0131'
PASS decode('latin5', '%80') is 'U+20AC'
PASS decode('latin5', '%9F') is 'U+0178'
PASS decode('latin5', '%FD') is 'U+0131'
PASS decode('windows-1254', '%80') is 'U+20AC'
PASS decode('ISO-8859-13', '%A1') is 'U+201D'
PASS decode('ISO-8859-13', '%A5') is 'U+201E'
PASS decode('ISO-8859-13', '%B4') is 'U+201C'
PASS decode('ISO-8859-13', '%FF') is 'U+2019'
PASS decode('windows-1257', '%80') is 'U+20AC'
PASS decode('windows-1257', '%B4') is 'U+00B4'
PASS decode('windows-1257', '%FF') is 'U+02D9'
PASS decode('iso-8859-7', '%A1') is 'U+2018'
PASS decode('iso-8859-7', '%B5') is 'U+0385'
PASS decode('iso-8859-7', '%B6') is 'U+0386'
PASS decode('windows-1253', '%80') is 'U+20AC'
PASS decode('windows-1253', '%A1') is 'U+0385'
PASS decode('windows-1253', '%B5') is 'U+00B5'
PASS decode('windows-1253', '%B6') is 'U+00B6'
PASS decode('KOI8-R', '%A4') is 'U+2553'
PASS decode('KOI8-R', '%AD') is 'U+255C'
PASS decode('KOI8-U', '%A4') is 'U+0454'
PASS decode('KOI8-U', '%AD') is 'U+0491'
PASS successfullyParsed is true

TEST COMPLETE
